low-level hack
This is supposed to be getting the amonut of bytes to read and the
address they are coming from without actually consuming them, which is
what we need for our implementation. However this always returns 0 on
macOS, haven't tested on linux.

This is based on https://github.com/dpifke/golang-wpasupplicant/blob/4b5b2dde8b5524b6737426ea25d4aa5373c3c458/unixgram.go#L125-L132