simplify protocols/protocolConn/...
Use higher level interface (writebuf/readbuf) which is all that is
needed in the switch.go code.
Then there are 'packet' connections which only send data, and 'stream'
connection which first send the packet size, and then the data.
This can be dealt with with 2 different types, packetConn and streamConn
with 2 different ReadBuf/WriteBuf implementation

Then the packetConn implementation reads/sends the size before the data,
and we also need 2 alternatives, 2 byte size (hyperkit) and 4 byte size
(qemu)