start adding high-level 'vm' type
The idea is to abstract the vfkit/qemu/… difference and to have a single
way of starting/stopping the VM and running commands / copying files in
it.

For now, this will only wrap SSH and allow sharing this code between
qemu and vfkit

FIXME: This apparently causes `make test-mac` to hang at the end of the tests?